feat: completed the modules task
